Wall Jumping/Sliding and Block Spawning
Player now slides on walls due to a high friction wall material and can jump/dash off of the wall.

Dash is now consumed when used off of a floor/wall instead of only being limited by a timer.

The player's horizontal movement stops when the player is not inputting horizontal movement and _limitVelocity is true to fit Hollow Knight's movement mechanics.

Created a prefab for a 1x1 Falling Block and a Block Spawner to randomly spawn normal and hazardous blocks with random fall speeds on a timer. Blocks are loaded from folders in Resources.

TODO: Change jump to match Hollow Knight mechanics by making the player go higher the longer it is held and start falling immediately when the button is released. Adjust movement speed, jump and dash strength, and velocity limits as feels best. Improve grounded/wall sliding detection.