Added Vertically Scrolling Camera + Background + Player Death
Camera now follows the player upwards.

Added temporary blue gradient background with sidescrolling clouds to make it easier to see that the player is gaining height.

Added health, isAlive, and heght values to PlayerController. It will now track how far the player has climbed/fallen and will kill the player (sets health to 0) and pause the game (sets Time.timeScale to 0) when the player falls below the screen.