Improved Player Movement and Starting Platform/Falling Blocks Destruction Logic
Changed player movement from setting velocity to apply a strong force and limiting velocity separately on the X and Y axis. X velocity is still set to 0 when there is no horizontal movement; movement is disabled when X velocity needs to not be set to 0.

Grounded and Wall Sliding now both use 3 raycasts for improved detection.

Falling blocks now destroy themselves 1 unit below the player's starting height instead of when below the screen, so that the player can fall onto blocks that have fallen below the screen but not below the player's death point.

When the player jumps for the first time, instead of destroying the starting platform, it now has a script on which the SwitchToDeathBox() method is called and it changes into a semi-transparent red color with a trigger that kills the player on contact.

This means the player now dies when touching the death box instead of when falling 1 unit below their starting point. This does make the game just slightly harder >:)